+++ lib/middleman-presentation/logger.rb
@@ -1,0 +1,8 @@
+# encoding: utf-8
+module Middleman
+  module Presentation
+    # Logger
+    class Logger < FeduxOrgStdlib::Logging::Logger
+    end
+  end
+end