+++ lib/middleman-presentation/transformers/ignore_slides.rb
@@ -12,6 +12,12 @@ module Middleman
 
         def initialize(ignore_file:, ignore_file_builder: IgnoreFile)
           @ignore_file = ignore_file_builder.new(ignore_file)
+
+          invalid_ignore_file = Pathname.new(ignore_file).dirname + Pathname.new('.slideignore')
+
+          if invalid_ignore_file.exist? 
+            Middleman::Presentation.logger.warn "Invalid ignore file \"#{invalid_ignore_file}\" detected. I'm going to ignore it. Please use the correct one \"#{ignore_file}\"." 
+          end
         end
 
         def transform(slides)